Add regex check to fix bug and extra check for nickname

+++ modules/reply.js
@@ -1,15 +1,13 @@
 const Discord = require("discord.js");
 const { prefix, customReactionPrefix } = require("./../config.json");
 
-
 module.exports = (client = Discord.Client) => {
 
   reply = async function reply(message) {
 
     if (message.channel.type === "dm" || message.author.bot || message.system) return;
     if (message.content.startsWith(prefix) || message.content.startsWith(customReactionPrefix)) return;
-
-
+    
     var content = message.content.split(/\r?\n/);
     const nameAndTime = content.shift();
     content = content.join('\n');
@@ -18,33 +16,51 @@ module.exports = (client = Discord.Client) => {
     members = members.map(m => {
       return {
         name: m.user.username,
+        nickname: m.nickname,
         dp: m.user.avatarURL
       }
     });
 
-    var fName;
-    var fDp;
-    var fTime;
-    var fChannel;
+    const regDate1 = /^Last (Monday|Tuesday|Wednesday|Thursday|Friday|Saturday|Sunday) at (1[0-2]|0?[1-9]):[0-5][0-9] (AM|PM)$/;
+    const regDate2 = /^(Yesterday|Today) at (1[0-2]|0?[1-9]):[0-5][0-9] (AM|PM)$/;
+    const regDate3 = /^\d{1,2}\/\d{1,2}\/\d{4}$/;
+
+    var fName, fDp, fTime, fChannel;
     var isFakeNews = true;
 
     for (m of members) {
-      if (nameAndTime.includes(m.name)) {
-        fName = m.name;
-        fDp = m.dp;
-        fTime = nameAndTime.slice(m.name.length);
 
+      const regName1 = new RegExp(`^${m.name}`);
+      const regName2 = new RegExp(`^${m.nickname}`);
+
+      if (regName1.test(nameAndTime) || regName2.test(nameAndTime)) {
+
+        var displayName;
+
+        if (regName1.test(nameAndTime)) displayName = m.name;
+        if (regName2.test(nameAndTime)) displayName = m.nickname;
+
+        fTime = nameAndTime.slice(displayName.length);
         if (fTime.includes('BOT'))
           fTime = fTime.slice(3);
-        break;
+
+        if (regDate1.test(fTime) || regDate2.test(fTime) || regDate3.test(fTime)) {
+          fName = displayName;
+          fDp = m.dp;
+          break;
+        }
       }
     }
 
-    if (!fName && !content) return;
+    if (!fName || !content.length || !fTime) return;
 
-    await message.delete();
+    message.delete();
 
-    const aName = message.author.username;
+    var aName = message.author.username;
+
+    const aNickname = await message.guild.fetchMember(message.author).then((u) => u.nickname);
+    if (aNickname) aName = aNickname;
+
     var toSend = content + `\n\n\`${fTime} | ${aName}\``;
 
     const msgEmbed = new Discord.RichEmbed()
@@ -64,16 +80,15 @@ module.exports = (client = Discord.Client) => {
           let fetched;
           let lastMsg = message;
 
-          //check if it can be retrieved in the last 500 messages in every text channel
           for (let x = 0; x < 5; x++) {
             fetched = await c.fetchMessages({ limit: 100, before: lastMsg.id });
             allMsg = await allMsg.concat(Array.from(fetched.values()));
             lastMsg = allMsg[allMsg.length - 1];
           }
 
-          allMsg = await allMsg.filter(msg => msg.author.username === fName);
-          await allMsg.reverse();
-          allMsg = await allMsg.join('\n');
+          allMsg = allMsg.filter(msg => msg.author.username === fName);
+          allMsg.reverse();
+          allMsg = allMsg.join('\n');
 
           if (allMsg.includes(content)) {
             fChannel = c.id;
@@ -94,6 +109,5 @@ module.exports = (client = Discord.Client) => {
         msgToEdit.edit(msgEmbed);
       })
     });
-
   };
 };
